feat: normalise game dimensions
Fixed dimensions regardless of browser window size is bad responsiveness
but should enable multiplayer to be consistent.
I should look into scaling as an alternative so that we can fill the
screen while keeping an aspect ratio. It should take a consistent amount
of time for both the ball to go across regardless of the scaling.